Overhauled ghost-stats script (#22987)
ref https://linear.app/ghost/issue/PROD-641/bundle-ghost-stats-script

We initially pulled the Tinybird web analytics tracking script, and due
to limitations in it, served our own copy of it with minimal
modification. Now that we're fully intend to use this script, we need to
update/rewrite it.

One of the major challenges is that we want parity between our
`member-attribution` url parsing (pulling out the ref, source, utm
params) and this script. We had copy-pasted that as well, which leads to
possible issues down the road with maintenance.

This PR adds a bundler step to serving the public asset prior to
minification so that we can use imports like `uuidv4` and
`tryghost/timezone-data`. This also pulls out the url attribution
functions into a utility file that can be used by the
`member-attribution` script in a separate commit to follow.

The script itself was overhauled in this process, pulling out all the
basic functions into utility files we can easily unit test. We moved to
using `fetch()` which is supported by any 'modern' browser (~past 5y).
If this proves difficult, we can move to a polyfill. We may want to add
logging to capture this when it does occur.