Fixed stale dependencies in docker compose volumes (#22996)
ref https://ghost.slack.com/archives/C08BZH42QF8/p1745318456373979

The docker compose setup includes named volumes for all the
`node_modules` directories in the monorepo, to avoid conflicts between
local dependencies and those installed in the container. This works for
the first run because the volumes are empty, so the dependencies
installed in the container at build time are copied into the volumes.

However, when the dependencies are changed, the stale dependencies
remain installed in the `node_modules` volumes, overwriting the
_correct_ dependencies that are installed in the container itself. This
leads to dependency mismatch issues, which can only be resolved by
deleting all the `node_modules` volumes. This can be done with `docker
compose down -v`, but this _also_ deletes the `mysql-data` and
`redis-data` volumes, which is inconvenient at best.

This commit adds a new package.json script, `docker:clean` that
selectively deletes all the `node_modules` volumes from the project. The
next time the project is run with `docker compose up`, the volumes are
recreated and the dependencies from the container are copied into them.
It also runs the `docker:clean` command in the `docker:build` script, so
this is done automatically whenever rebuilding the container, provided
you use the `yarn docker:build` command.