Reverted bundled frontend scripts changes (#23005)
ref https://ghost.slack.com/archives/C07HTEJMR2R/p1745413896875469

Oddly, this passed just fine, but eventually failed in later commits
with the Canary build with a clear issue in the path: `EACCES:
permission denied, open
‘/home/ghost/core/frontend/src/ghost-stats/ghost-stats.bundled.js’`
(should be /core/core/).

The minified is what should read the bundler file. At any rate, it's
been called to my attention that there's an existing issue
(https://github.com/TryGhost/Ghost/pull/21866) to move these
runtime-minified assets to build-time minification. The reason some of
these assets (like card-assets) were built at runtime is so they can be
rebuilt when applying new themes, as they can be dependent on them.

The best approach here is to resolve the open PR for build-time
minification, then re-apply these changes in spirit with that same
approach to the ghost-stats and member-attribution scripts.