Refactored boot time / on demand asset generation (#23006)
ref
[ONC-662](https://linear.app/ghost/issue/ONC-662/fix-file-write-issues-in-ghost-application-related-to-asset-generation)
ref
[5705c6c](https://github.com/TryGhost/Ghost/commit/5705c6cfa1c0d73fe364cf75170d9b059d235bd0)

Refactored the generation of the following assets:
ghost-stats.js
comment-counts.js
member-attribution.js
admin-auth
to occur at build time instead of at boot time or on demand

Cards are still built on demand as these assets need to be re-generated
when a theme changes - we utilise the existing asset minification system
for this, except we only execute the middleware responsible for this
when a request is for a card asset (instead of any path, which is a bug
in the current implementation).

This also reverts the changes made in
https://github.com/TryGhost/Ghost/pull/21857 as these changes are no
longer needed (because the assets are now generated at build time).

Co-authored by @mike182uk. Copied the above from #21866 and updated as
it's been a few months, plus added a few changes to accommodate the
needs of `ghost-stats` which will come in a subsequent commit.