Suppressed console logs during JS asset minification (#23010)
ref
https://github.com/TryGhost/Ghost/commit/995e119ae67679610cc0a5679331834d7ca09503

The `minify-assets.js` had some console logs that were running on boot.
These logs are useful for debugging, but most of the time it should be
fine to do the asset minification silently. This commit downgrades the
`console.log()`s to `logging.debug()`s, so we can still enable it if
we're debugging the asset minification, but it doesn't add noise to the
console.

It also adds an NX `depends_on` configuration for `yarn test:all` (which
includes unit tests and was failing unless you had manually run
`build:assets` or run `yarn dev` previously) and the `yarn
test:browser*` commands to ensure the assets are built when running
browser tests.