🎨 Improved handling of comments for themes with darkmode toggles (#22407)
Closes https://github.com/TryGhost/Ghost/issues/21103

"Auto" for comments works great for themes that detect browser
preferences or theme settings. However, themes with toggle switches can
end up looking rather strange when the toggle is flipped.

This PR adds a mutation observer that watches the parent of
#ghost-comments-root for class or style changes. If either is detected,
it re-runs the darkmode function and updates new state, triggering a
re-render in the correct color scheme.

Theme creators with toggle switches can trigger the behavior either by
adding a class to the parent of #ghost-comments-root or by directly
adding a color to that element's style. Note that mutation observers do
not detect when CSS variables or inherited colors change, so other
change types (inherited colors from an ancestor or changes to a
variable) will not trigger an update.

I considered detecting changes at body or html, but some themes make a
lot of unrelated changes to those elements, so it seemed safest to stay
local, at the risk of not "just working". I considered some heavier
packages that attempt to observe all style changes, but wasn't convinced
that these workarounds were both robust and lightweight.

---------

Co-authored-by: Steve Larson <9larsons@gmail.com>