Prevented database tests running in CI when only Tinybird files have changed (#23073)
no refs

To make CI run faster on PRs that only change Tinybird datafiles, we
shouldn't run the database tests if only Tinybird datafiles have
changed. This had been working with the Tinybird files in
`ghost/web-analytics`, but now that the Tinybird FORWARD files live in
`ghost/core`, it was triggering the database tests. This commit adds an
exclusion, so it will ignore changes to
`ghost/core/core/server/data/tinybird` when deciding whether or not to
run the database tests.