Removed `.devcontainer` directory (#23075)
ref https://ghost.slack.com/archives/C02G9E68C/p1745367891427209

This `.devcontainer` config was from an initial attempt at a Dockerized
development environment. We ultimately found that the Dev Container
specific configuration made it more difficult to debug when things went
wrong, as there it adds a semi-opaque layer of orchestration on top of
the basic docker compose setup.

We've since moved toward using docker compose directly for local Docker
development, and it's working much more smoothly.

This `.devcontainer` directory therefore is no longer used/needed, so
this commit cleans it up.