Fixed CI skipping `admin-x-settings` tests if only `admin-x-framework` changed (#23077)
no refs

The `admin-x-settings` packages all depend on `admin-x-framework` in
some capacity, so we should always run their tests if
`admin-x-framework` changes. This wasn't happening previously — this
commit fixes that.

Without this, it's possible to make changes to `admin-x-framework` which
inadvertently break `admin-x-settings` without realizing it. With this
fix, we'll always run the tests if we update the framework, which should
catch such instances.