Fixed filtering by post in post analytics web stats (#23076)
ref
https://linear.app/ghost/issue/PROD-1541/add-a-filter-by-post-uuid-to-all-our-pipes

The Post Analytics web stats view was filtering to the post using the
`pathname` parameter, based on the slug of the post. This doesn't work
for posts that are impacted by custom routing, since the events have the
full pathname, while the filter we were using here only includes the
slug.

This commit updates this view to filter on the `post_uuid` instead of
the pathname, so we can accurately return all the pageviews for the
post, regardless of custom routing or if the slug ever changes.