Added `post_uuid` to top pages pipe (#23096)
ref https://linear.app/ghost/issue/PROD-12/

The top pages pipe/endpoint currently only outputs the pathname and
count. We want to be able to display the name of the content (post name,
page name, etc), so we need to expose the post_uuid.

Because we need to mesh this data with Ghost data, we'll have to follow
this up with Ghost's backend doing the fetching from Tinybird and
meshing that with data from Ghost's db.