Added `jsonc` parsing for local config (#23097)
no refs

It would be nice to be able to comment out parts of our local
configuration when in development, so we can easily enable/disable
certain blocks of configuration depending on what we're working on
without having to completely delete it. Currently we can't do this
because our config loader only accepts `.json` files, and the standard
JSON spec does not allow comments.

However, nconf can be extended to use custom file formats, and that's
exactly what this PR does. With this current implementation, Ghost will
look for a `ghost/core/config.local.jsonc` file in your local repo, in
addition to the default `config.local.json` file. If you rename your
current `config.local.json` to `config.local.jsonc`, it will be loaded
instead, and you can enable/disable blocks of config using JavaScript
style comments.

The actual `jsonc` parsing is handled by
`[jsonc-parser](https://www.npmjs.com/package/jsonc-parser)`, which is a
small node project maintained by Microsoft.

There may also be applications for doing this in other configuration
files for the purpose of documentation, but for now this only adds
`jsonc` parsing for the `config.local.jsonc` file.