Updated stats content to use names instead of paths (#23098)
ref https://linear.app/ghost/issue/PROD-12/
- created stats api endpoint in admin-x-framework to be used by stats
views
- added new /top-content endpoint that pulls data from tinybird and
ghost
- added tinybird client to stats service for fetching tinybird data
- added names of content to the content block in the stats>web view
- moved config use to backend from frontend; reduced query parameters
passed/exposed in request

UX can be improved by displaying the content name (post title, page
name, author's name from author page, 'Home' page) in the Content block
in the Web tab of Stats. This provided some new challenges that we'll
need to make more use of, specifically meshing Tinybird data with Ghost
data.

The Tinybird pipes/endpoints were updated previously to expose the
post_uuid for the top_pages endpoint. This is now requested by Ghost
(the backend) instead of the frontend, where the post_uuids are looked
up in the database for the post title. Where we don't have a post_uuid,
we use the resource lookup provided by the urlService.