Added `split` profile to docker compose to run ghost, admin, apps separately (#23114)
no refs

This commit splits the current `ghost` all-in-one docker compose service
into individual services for all of the different apps and processes
that are currently run in `dev.js`. Currently these are all behind the
`split` profile, but after working through the kinks over the next ~week
we'll probably delete the current `ghost` all-in-one service in favor of
this setup.

## Why? I'm glad you asked.

The current configuration with the `ghost` all-in-one service simply
loads the whole monorepo into the container and runs the `dev.js`
script, which is how we run Ghost and all its supporting apps locally.
This has proven to be problematic for a few reasons, and it's also
_very_ non-idiomatic — it makes sense when developing locally, but it
makes a lot less sense in a docker compose based setup, where we should
be leaning into docker compose to do this orchestration for us, rather
than manually running a dozen subprocesses in a bespoke script.

The `dev.js` script creates a few problems in Docker:
1. It relies on the NX Daemon because it uses `nx watch` to watch &
rebuild all the packages that have a build step. The NX Daemon can only
run in one instance of the container, so trying to run ghost & run tests
in separate containers causes the daemon to crash.
2. Some of the commands in `dev.js` don't always exit cleanly, leading
to _more_ issues with the NX Daemon. This leaves the container in a
weird state and causes it to hang on startup after the initial run. This
is ultimately why we have the `--force-recreate` flag in our `yarn
docker:dev` command, so we always start with a brand new container. This
workaround has been working, but it means we can't really use the bare
`docker compose` commands.

## How does this commit fix those problems?

There are two key changes that fix these problems:

1. By running all the different processes in their own services instead
of in a dozen subprocesses, we obviate the need for the NX Daemon in
these services. Most of the apps have their own file watching built into
their `yarn dev` command for hot reloading, so it's possible to run each
app individually without the daemon being enabled. This alone fixes the
issues with the NX Daemon by essentially deleting it.

2. But then there's a problem — each app's `yarn dev` command can watch
its own files and reload on changes, _but what if one of its
dependencies in the monorepo changes?_ This is where the standalone
`daemon` service comes in. This service is the only one that has the NX
Daemon enabled, and it only has one job: watch all the packages in the
monorepo that have a `build` target, and rebuild them when they change.
Since all the different services share the same bind mount, the `daemon`
service can rebuild e.g. a typescript dependency of the `server`
service, and the `server` service will see that its dependencies have
changed, and reload itself — like ✨magic✨.

Putting the two together, we now have all the nice conveniences of the
`dev.js` script in a much more "docker first" paradigm.

## Are there any other benefits to these changes?

Absolutely! These other bits weren't the main motivation for this
change, but this commit does address a few key problems that previously
required annoying workarounds to solve:

1. As mentioned, with this profile enabled, you can now just run `docker
compose up` and everything should work just fine. No more NX Daemon
issues means no more `--force-recreate` is needed.

2. Also an unintended but happy consequence of this change: everything
_feels_ much faster. I think the explanation for this is twofold:
running each process in its own container is generally more performant
than delegating to subprocesses, and running fewer file watching
processes reduces the number of re-crawls. In particular, the admin
ember build feels much faster and I haven't yet seen all the
`recrawling` warnings that are commonplace when running `dev.js` locally
or in docker.

## Further opportunities
This establishes a solid pattern that should allow us to start getting
really creative with our docker compose setup going forward. Now that we
can run a bunch of mostly identical services in parallel without causing
problems, we can do lots of cool shit:

1. Throw a reverse proxy like `nginx` in front of all these services so
we can use custom hostnames for each service, make `https` the default,
configure subdirectory sites, etc. With this, we can also probably
consolidate some of these additional services. All the "apps" are just
static files, so we should be able to have the `daemon` service handle
watching and rebuilding these, and offload all the serving to a single
`nginx` service that just serves the built files.

2. Cleaning up the output. This is more of a "todo" than an opportunity,
but just wanted to acknowledge that the console is a bit noisy with some
of these changes, which we can and should work to reduce so it doesn't
feel so overwhelming and difficult to trace what's happening.

3. Create a dedicated service for each test suite, slap a `test` profile
on them, and run all our test suites in parallel, which I'm hopeful will
be way faster than CI is currently.

4. Improve our docker build caching. Currently since everything is
running in the same service, there are some additional (painfully slow)
steps in the Docker build process that are only needed for specific
use-cases. Now we can move those bits into their own build targets, and
only use them when we need to. The most obvious opportunity here is
playwright — it's currently installed on every build, but we only need
it when running browser tests. This will let us create a dedicated build
target that includes "ghost + playwright", and this target would only be
used by a dedicated `browser-tests` service. There's similar
opportunities with e.g. the Tinybird CLI, which could also have a
dedicated service now.

5. Create a really awesome development environment in the `daemon`
service, which we can attach our IDEs to for development. We can create
a new build target that extends the `development` target to include all
kinds of bells and whistles — shell prompts, aliases, git configuration,
ssh agent forwarding, extra unix packages that we don't want to
introduce into the main `development` target, default VSCode extensions,
etc.

---------

Co-authored-by: Hannah Wolfe <github.erisds@gmail.com>