Fixed the ability to overwrite protected settings key with data import (#23136)
closes ONC-900

When importing a data file that included updating the `email_verification_required` flag, the key would successfully be updated because the setting is not part of the `ignoredSettings` list. This can prevent our email verification logic from working and open the door for malicious spammers.

Adding this key to the list of ignored settings gives basic protection from intentionally or accidentally overwriting this value with an import.