Reverted card assets generation mw changes (#23137)
ref https://ghost.slack.com/archives/C025584CA/p1746099965766049
ref https://github.com/TryGhost/Ghost/pull/23006/

We had a handful of reports of missing/empty cards.min.css files. We
made recent changes to the middleware to only attempt to generate on
request for the assets rather than every request, and this may have
exposed gaps in the minification process' fidelity. We never had reports
of this previously, so it seems worth reverting this portion.