Wired up member counts to stats growth page (#23139)
ref https://linear.app/ghost/issue/PROD-658/

The growth tab currently has fixture data for demo purposes. This wires
up the member counts endpoint used by the Admin Dashboard. We had to
make a small adjustment to allow for a start date, as it was previously
always looking back 91 days. We will probably want to make some
additional adjustments to this down the road.