Added dedicated tinybird target to improve build time (#23153)
no refs

The Tinybird CLI installation in the Docker build was running after all
the code and node dependencies had been copied in and installed. This
means if the build cache was busted by a dependency change, we'd also
have to reinstall the tinybird CLI, even if nothing had changed.

We only need the Tinybird CLI installed in select circumstances, so this
commit creates a new target in the Docker build that inherits from the
base development image and just installs the Tinybird CLI. This should
speed up builds and rebuilds from the main docker build and for
subsequent rebuilds.