Added dedicated browser-tests target to improve docker build time (#23154)
no refs

Currently playwright is installed as part of the base development build.
The playwright install step needs to be after the main `yarn install`
step to ensure we install the correct version, and as a consequence, we
have to reinstall playwright everytime any dependency or package.json
changes.

This commit gets around this problem by creating a dedicated build
target for browser-tests that includes the playwright build. The `yarn
docker:test:browser` command uses this build target, while the regular
`yarn docker:dev` command uses the development target, so now we only
have to wait for playwright to install if we're actually running the
browser tests.