Removed job queue manager class
ref https://linear.app/ghost/issue/ENG-2028/

We created this persisted queue using a MySQL-based queue to prototype a more robust queuing system, intending to use Redis, RabbitMQ, or the like. We didn't find enough benefit from this and found other constraints, primarily r/w to the db being burdensome, and not enough benefit to justify the move to RMQ/Redis at this time.

Given that, we're removing the job queue manager. We've previously removed any calling code, so now we're pulling the class itself and will follow up with a migration (or two) to clean up the jobs table schema and entries.