Updated member upgrade to use welcome page url if provided (#23173)
ref https://github.com/TryGhost/Ghost/issues/23095

We have the concept of a welcome url for tiers, which we utilize for new
signups via `createSessionFromToken`. We do not use this code path for
already-authenticated users, the result being that the paid tier URL is
not used whenever an existing member upgrades their tier (or
downgrades).

This certainly isn't the best supported feature out there. I've updated
the router the handle this upgrade case, although there's likely better
ways about it.