Reverted admin auth middleware & minification changes (#23179)
ref https://github.com/TryGhost/Ghost/pull/23006
ref https://linear.app/ghost/issue/ONC-919/

Following the minification changes to the various frontend public
assets, it seems the auth frame is broken on Ghost Pro. It's unclear why
this is - it still works perfectly fine locally, and I wasn't able to
easily track down the cause given the urgency vs. benefit. Frankly,
there's not much benefit here besides trying to establish a new model
that a bit more purposeful.

It seems likely it was an issue in the middleware and not the
minification process.