Added mutations for blocking & unblocking domains
ref https://linear.app/ghost/issue/PROD-697

These allow the domain level block feature to start being wired up to the API. We introduce a new property domainBlockedByMe to the account, which can be used to determine if an account is blocked at the domain level. It is currently not persisted, it's only in memory of the current session.