Added Post Analytics Growth totals (#23185)
ref
https://linear.app/ghost/issue/PROD-1542/adjust-conversion-attribution-to-reflect-free-members-and-paid-members

On the Post Analytics Growth tab, we want to show total free and paid
members + MRR attributed to the post. I'd initially planned on just
summing the result of the `stats/posts/:id/top-referrers/` response on
the frontend, but there is a limit on this endpoint so that wouldn't
work.

This commit adds a new endpoint explicitly for calculating the totals,
and wires it up to the frontend view.