Removed alpha flags from Stats and Post Analytics Growth tabs (#23197)
ref
https://linear.app/ghost/issue/PROD-657/add-growth-tab-to-sitewide-context
ref
https://linear.app/ghost/issue/PROD-658/add-growth-tab-to-post-analytics-context

We've been working on these Growth tabs behind an additional alpha flag
so we can iterate on them without releasing them to the Beta just yet.
We're now getting ready to ship these new tabs to the Beta, so this
commit removes the conditionals on the `trafficAnalyticsAlpha` flag,
which will show these tabs to anyone with the beta enabled.