Updated release-apps script (#23223)
no ref
- bumped announcement bar
- added announcement-bar to the new-ish release-apps script
- updated release-apps script to not dump empty commits when checking
the last 50

This contains changes to the `release-apps` script. This *may have some
problems* as it's a little difficult to test without making additional
changes to the apps. I will try to do this next time there's changes,
the problem currently is that I just shipped all of them, and the script
only looks at the last 50 commits. Announcement-bar hasn't been shipped
in a long time so we don't expect any to fall in that window.

The reason for the change to the script is that it was dumping 50
entries of links to commits w/o a hash (commits that didn't actually
touch the repo), which was erroneous and bothersome.