Improved tests in the `stats` app (#23226)
no refs

The tests in the `stats` app weren't running in CI, and a few of the
View tests were failing. This commit adds the `yarn test:unit` command
to the package.json, which will run the unit tests in CI.

It also deletes the existing View tests for now. These types of tests
seem to be better suited to run in Playwright, where we can run the
whole app. This avoids a lot of the boilerplate mocking setup for each
View test, which makes the tests harder to write and reason about.