Improved the Dockerfile for faster incremental builds (#23229)
no refs

This commit makes a number of improvements to the monorepo's development
Dockerfile, with an overall goal of improving the build time and build
caching.

Here are the highlights:
- Moves the Playwright install to the top of the file, so this step will
be cached unless the playwright version we're using changes. Previously
the playwright install ran after everything else, which means we had to
reinstall playwright if any code or package.json changed.
- Adds a cache mount for the `yarn install` step. This reduces the `yarn
install` time on warm builds from ~90 seconds to ~30 seconds from my
testing, by caching the downloaded packages in a volume that is managed
by Docker.
- It also adds a few additional directories to the `.dockerignore` file,
which significantly reduced the size of the final image by ~50%