Changed flag checks to maintain path in Stats
no ref

The stats page would always push the user to `/stats/web/` on refresh because it was loading the alpha flag. This is annoying. I've instead introduced a higher order component to wrap each view that allows us to set a redirect if needed. Now we'll wait for the flag before loading the tab instead of pushing the user to `/web/` and still not loading it.