Fixed browser tests running locally in docker (#23243)
no refs

The `yarn docker:test:browser` command was broken when the
`browser-tests` target was removed from the Dockerfile. This updates the
command so it works without the `browser-tests` target. It also switches
back to using `docker run ...` to run the browser tests, rather than the
`browser-tests` profile in the compose file.