Wired up Post Analytics > Newsletter stats (#23249)
ref
https://linear.app/ghost/issue/PROD-1648/post-analytics-calculate-average-open-and-click-rates-for-the-average

On the new Post Analytics > Newsletters tab, we want to show some top
level metrics for the current post (sent, opened, clicked), plus a
comparison of this newsletter to the "average" newsletter.

This commit wires up the data to the static views that Zimo created.
This doesn't yet include the list of links at the bottom of the view,
which I'll follow up with in a separate PR.