Fixed Post Analytics > Newsletter tab premature redirect (#23270)
ref
https://linear.app/ghost/issue/PROD-1672/post-analytics-newsletter-tab-fails-to-load-frequently

The Newsletter tab in the Post Analytics view was prematurely
redirecting because the `trafficAnalyticsAlpha` flag hadn't been loaded
yet. The redirect was occurring before the settings were loaded from the
server, so unless the `posts` app had already been loaded and cached the
settings, the Newsletter tab would redirect to the `/web/` route. This
`/web/` route doesn't really exist, so it shows the "Loading
interrupted" error.

This follows the same pattern we have in the `stats` app using a
`withFeatureFlag` higher order component to put the whole route behind
the flag, and critically it waits for the settings to load before
redirecting.

Note: this duplicates the `withFeatureFlag` and `useFeatureFlag` hook
from `stats` into `posts`. We should either merge these apps together,
or move these utilities into `admin-x-framework` to remove the duplication.