🐛 Fixed comments UI showing logged-in member's avatar for comments by a deleted member (#23304)
ref
https://linear.app/ghost/issue/PROD-599/in-comments-the-deleted-member-avatar-is-replaced-with-the-avatar-of

The avatar displayed for a comment made by a deleted member was
defaulting to the currently logged in member's Avatar, which is
confusing and incorrect. The Avatar component accepted either a comment
or a member, and it would fallback to displaying the logged-in member's
avatar if `comment.member` wasn't truthy, as is the case for a deleted
member where `comment.member` is `null`.

This commit simplifies the Avatar component to only accept a member
prop, rather than accepting either a member or a comment prop. This
simplifies the logic in the Avatar component, which can now just render
the avatar for the member provided, or a blank avatar if the provided
member is `null`, without any fallback logic. It's currently only used
in two places, so it's easy enough to update the uses where it was
previously provided a comment.