🐛 Fixed new paid member signup error for sites with many newsletters (#23303)
ref
https://linear.app/ghost/issue/ONC-939/support-escalation-bytetree-members-cannot-checkout-via-offer

Portal sends newsletter names and ids in the request it makes to
`/members/api/create-stripe-checkout-session/` when a new member signs
up for a paid account. Ghost then adds the list of newsletters to the
checkout session's `metadata` field, including the names and ids.

For sites with enough newsletters, this can exceed the character limit
of the checkout session's `metadata` field. This commit changes Ghost's
handling so it only adds the newsletter `ids` to the `metadata` field,
which gives sites with a lot of newsletters a little more breathing room
to stay under this limit. If the limit is reached, Stripe will return a
400 error, preventing new members from starting paid accounts.