🐛 Updated EmailRenderer+css to use serif headers if serif body selected
fixes [issue #2205](https://github.com/TryGhost/Ghost/issues/22053)
Headers (h1-h6) in email newsletters are currently shown in sans-serif font, despite the fact that users can select serif fonts for the body of their newsletter. This is because the sans-serif default css is not overriden for header dom elements. This PR takes a similar approach to overriding the default css as is used with the excerpt font, so that email font overrides are all created in the same place and in the same manner.