🐛 Fixed CTA for public preview card not showing on post previews (#23182)
closes https://github.com/TryGhost/Ghost/issues/23181
closes
https://linear.app/ghost/issue/ONC-921/oss-issue-content-cta-does-not-display-on-preview-when-it-should
refs
https://linear.app/ghost/issue/ONC-930/support-escalation-re-freepaywalls-on-web-versions-of-email-only-posts

When a post is set to members only or paid only and there is a public
preview card present, the "This post is for members only" CTA wasn't
appearing in the post preview, even though the private content was
correctly omitted. The CTA was correctly being displayed on the actual
post on the frontend, however.

This pattern of setting `post.access` in the frontend controller route
was previously used in all of the `entry`, `preview` and `email`
routers. This
[commit](https://github.com/TryGhost/Ghost/commit/fa91c6c9545a9107ba9b938f545c674ff70a462c#diff-42f164398878ef21b6e9ab8e68f273a167fe1f5e85f4b3fe0036d92ab8128e94)
from 2020 switches this pattern to calculate the value of `post.access`
in the content API rather than the frontend, and this same `post.access
= !!post.html` line was removed from the `entry` controller at that
time. However, the change wasn't carried over to the `email` or
`preview` controllers, where the `post.access` value from the API
was being overwritten.

This commit carries the pattern over to the `preview` controller, and
adds a test to ensure the CTA is rendered in the preview route. I'll
follow up with another PR to do the same for the `email` controller.