Removed "restart mode" from `startGhost` test utils (#23348)
no refs

In our e2e test utils, the `startGhost` function has two modes: "Restart
mode" and "Fresh mode". The "Restart mode" is run when Ghost is already
running, and "Fresh mode" is used when it's not, or when the
`forceStart` option is passed.

These test utils have been moved around quite a bit so I haven't been
able to unearth the full history of why these two versions of the
`startGhost` util were created in the first place, but my assumption is
that it was intended as a performance improvement — why fully restart
Ghost if we can just reset the DB and some other internal state instead?

However:
1. This leads to unpredictable behavior in our tests that is a huge pain
to debug — if Test A starts Ghost and changes some internal state that
isn't reset by the "Restart mode", and Test B just starts Ghost, the
server could be in an unexpected state, leading to test errors. A good
example of this currently in the codebase is if you add a `return
testUtils.stopGhost();` line in this
[`email_routes.test.js`](https://github.com/TryGhost/Ghost/blob/5b692d21c17659c4b63b3162a7d2cdf41e04b923/ghost/core/test/e2e-frontend/email_routes.test.js#L37-L39)
after hook to stop the server, and run `yarn test:e2e`, now you'll see
test failures in `e2e-frontend/members.test.js`. Adding the `forceStart`
in the members test file fixes the problem, but:

2. From my own testing, defaulting to the "restart mode" only saves us
~1 second total when running database tests in CI, or running `yarn
test:all` locally in `ghost/core`, which is <1% performance improvement.

This PR removes the "restart mode", and always uses "fresh start mode" —
this means whenever you call `testUtils.startGhost()`, it will first
stop the Ghost server if it's running, before restarting. It also
removes the `forceStart` option and all instances of it, since it
doesn't change any behavior anymore, and renames the
`freshModeGhostStart` internal function to simply `_startGhost`, since
there's now only 1 version.