Updated variable names in `linkSubscription` for better readability (#23298)
no refs

The `linkSubscription` model in the members service is notoriously
complex and difficult to follow. This commit is a first step in
improving the readability of this critically important method of the
MemberRepository, to make it easier to follow the logic. There shouldn't
be any functionality changes here (hence no new/updated tests) — only
the locally scoped variables have been changed, mostly to be more
specific and to disambiguate between e.g. the stripe subscription object
from Stripe and the `stripe-customer-subscription` bookshelf model.